Move all ValidationProblemId classes to model-core
