+++ subprojects/core/src/integTest/groovy/org/gradle/api/tasks/DirectorySensitivityErrorHandlingIntegrationSpec.groovy
@@ -16,7 +16,7 @@
 
 package org.gradle.api.tasks
 
-import org.gradle.api.internal.validation.CoreValidationProblemId
+import org.gradle.model.internal.reflect.problems.CoreValidationProblemId
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.internal.reflect.validation.ValidationMessageChecker
 import org.gradle.internal.reflect.validation.ValidationTestFor
