+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/properties/annotations/TypeAnnotationHandlerSupport.java
@@ -15,7 +15,7 @@
  */
 package org.gradle.api.internal.tasks.properties.annotations;
 
-import org.gradle.api.internal.validation.CoreValidationProblemId;
+import org.gradle.model.internal.reflect.problems.CoreValidationProblemId;
 import org.gradle.internal.reflect.validation.TypeValidationContext;
 
 import java.lang.annotation.Annotation;
