+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/ValidateStep.java
@@ -28,7 +28,7 @@ import org.gradle.internal.execution.WorkValidationException;
 import org.gradle.internal.execution.history.BeforeExecutionState;
 import org.gradle.internal.execution.history.ExecutionHistoryStore;
 import org.gradle.internal.execution.history.PreviousExecutionState;
-import org.gradle.internal.execution.validation.ExecutionValidationProblemId;
+import org.gradle.model.internal.reflect.problems.ExecutionValidationProblemId;
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
 import org.gradle.internal.reflect.validation.Severity;
 import org.gradle.internal.reflect.validation.TypeValidationContext;
