+++ subprojects/model-core/src/testFixtures/groovy/org/gradle/internal/reflect/validation/ValidationMessageChecker.groovy
@@ -17,13 +17,13 @@
 package org.gradle.internal.reflect.validation
 
 import org.gradle.api.internal.DocumentationRegistry
-import org.gradle.api.internal.artifacts.validation.ArtifactValidationProblemId
-import org.gradle.api.internal.validation.CoreValidationProblemId
 import org.gradle.integtests.fixtures.executer.GradleExecuter
-import org.gradle.integtests.fixtures.validation.TestValidationProblemId
-import org.gradle.internal.execution.validation.ExecutionValidationProblemId
 import org.gradle.internal.reflect.JavaReflectionUtil
-import org.gradle.internal.reflect.annotations.TypeMetadataValidationProblemId
+import org.gradle.model.internal.reflect.problems.ArtifactValidationProblemId
+import org.gradle.model.internal.reflect.problems.CoreValidationProblemId
+import org.gradle.model.internal.reflect.problems.ExecutionValidationProblemId
+import org.gradle.model.internal.reflect.problems.TestValidationProblemId
+import org.gradle.model.internal.reflect.problems.TypeMetadataValidationProblemId
 
 import static org.gradle.internal.reflect.validation.TypeValidationProblemRenderer.convertToSingleLine
